@@ -1,6 +1,6 @@
 {
   "name": "laravel-file-manager",
-  "version": "2.5.4",
+  "version": "3.0.1",
   "description": "File manager for Laravel",
   "keywords": [
     "laravel",
@@ -12,33 +12,31 @@
   "license": "MIT",
   "main": "src/init.js",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "test:unit": "vue-cli-service test:unit"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "axios": "^0.21.1",
-    "codemirror": "^5.59.2",
-    "core-js": "^3.9.0",
-    "cropperjs": "^1.5.11",
-    "plyr": "^3.6.4",
-    "vue": "^2.6.12",
-    "vue-codemirror": "^4.0.6",
-    "vuex": "^3.6.2"
+    "axios": "^0.25.0",
+    "codemirror": "^5.65.1",
+    "codemirror-editor-vue3": "^1.0.1",
+    "cropperjs": "^1.5.12",
+    "mitt": "^3.0.0",
+    "plyr": "^3.6.12",
+    "vue": "^3.2.25",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.4.6",
-    "@vue/cli-plugin-eslint": "~4.4.6",
-    "@vue/cli-plugin-vuex": "~4.4.6",
-    "@vue/cli-service": "~4.4.6",
-    "@vue/eslint-config-airbnb": "^5.0.2",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.8.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "node-sass": "^4.14.1",
-    "sass-loader": "^8.0.2",
-    "vue-template-compiler": "^2.6.12"
+    "@vitejs/plugin-vue": "^2.0.0",
+    "eslint": "^8.8.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-vue": "^8.4.0",
+    "prettier": "^2.5.1",
+    "sass": "^1.49.7",
+    "vite": "^2.7.2",
+    "vite-plugin-environment": "^1.1.0"
   }
 }
