@@ -1,10 +1,10 @@
 <template>
     <div class="fm-content d-flex flex-column">
-        <disk-list v-bind:manager="manager"/>
-        <breadcrumb v-bind:manager="manager"/>
+        <disk-list v-bind:manager="manager" />
+        <bread-crumb v-bind:manager="manager" />
         <div class="fm-content-body">
-            <table-view v-if="viewType === 'table'" v-bind:manager="manager"/>
-            <grid-view v-else v-bind:manager="manager"/>
+            <table-view v-if="viewType === 'table'" v-bind:manager="manager" />
+            <grid-view v-else v-bind:manager="manager" />
         </div>
     </div>
 </template>
@@ -12,40 +12,39 @@
 <script>
 // Components
 import DiskList from './DiskList.vue';
-import Breadcrumb from './Breadcrumb.vue';
+import BreadCrumb from './BreadCrumb.vue';
 import TableView from './TableView.vue';
 import GridView from './GridView.vue';
 
 export default {
-  name: 'Manager',
-  components: {
-    DiskList,
-    Breadcrumb,
-    TableView,
-    GridView,
-  },
-  props: {
-    manager: { type: String, required: true },
-  },
-  computed: {
-    /**
-     * view type - grid or table
-     * @returns {default.computed.viewType|(function())|string}
-     */
-    viewType() {
-      return this.$store.state.fm[this.manager].viewType;
+    name: 'Manager',
+    components: {
+        DiskList,
+        BreadCrumb,
+        TableView,
+        GridView,
     },
-  },
+    props: {
+        manager: { type: String, required: true },
+    },
+    computed: {
+        /**
+         * view type - grid or table
+         * @returns {any}
+         */
+        viewType() {
+            return this.$store.state.fm[this.manager].viewType;
+        },
+    },
 };
 </script>
 
 <style lang="scss">
-    .fm-content {
-        height: 100%;
-        padding-left: 1rem;
+.fm-content {
+    padding-left: 1rem;
 
-        .fm-content-body {
-            overflow: auto;
-        }
+    .fm-content-body {
+        overflow: auto;
     }
+}
 </style>
