@@ -2,13 +2,11 @@
     <div class="modal-content fm-modal-about">
         <div class="modal-header">
             <h5 class="modal-title">{{ lang.modal.about.title }}</h5>
-            <button type="button" class="close" aria-label="Close" v-on:click="hideModal">
-                <span aria-hidden="true">&times;</span>
-            </button>
+            <button type="button" class="btn-close" aria-label="Close" v-on:click="hideModal"></button>
         </div>
         <div class="modal-body">
             <strong>{{ lang.modal.about.name }}</strong>
-            <hr>
+            <hr />
             <dl class="row">
                 <dt class="col-3">{{ lang.modal.about.version }}:</dt>
                 <dd class="col-9">{{ version }}</dd>
@@ -21,7 +19,7 @@
                     <a href="https://github.com/alexusmai/laravel-file-manager" target="_blank">
                         alexusmai/laravel-file-manager
                     </a>
-                    <br>
+                    <br />
                     <a href="https://github.com/alexusmai/vue-laravel-file-manager" target="_blank">
                         alexusmai/vue-laravel-file-manager
                     </a>
@@ -36,16 +34,16 @@ import modal from '../mixins/modal';
 import translate from '../../../mixins/translate';
 
 export default {
-  name: 'About',
-  mixins: [modal, translate],
-  computed: {
-    /**
-     * App version
-     * @returns {*}
-     */
-    version() {
-      return this.$store.state.fm.settings.version;
+    name: 'AboutModal',
+    mixins: [modal, translate],
+    computed: {
+        /**
+         * App version
+         * @returns {*}
+         */
+        version() {
+            return this.$store.state.fm.settings.version;
+        },
     },
-  },
 };
 </script>
