@@ -1,28 +1,28 @@
 export default {
-  /**
-   * Base URL
-   * @param state
-   * @returns {default.baseUrl|(function(*))|string|*|string|null}
-   */
-  baseUrl(state) {
-    return state.baseUrl;
-  },
+    /**
+     * Base URL
+     * @param state
+     * @returns {null}
+     */
+    baseUrl(state) {
+        return state.baseUrl;
+    },
 
-  /**
-   * Headers
-   * @param state
-   * @return {*}
-   */
-  headers(state) {
-    return state.headers;
-  },
+    /**
+     * Headers
+     * @param state
+     * @return {*}
+     */
+    headers(state) {
+        return state.headers;
+    },
 
-  /**
-   * Headers has Authorization
-   * @param state
-   * @return {boolean}
-   */
-  authHeader(state) {
-    return Object.prototype.hasOwnProperty.call(state.headers, 'Authorization');
-  },
+    /**
+     * Headers has Authorization
+     * @param state
+     * @return {boolean}
+     */
+    authHeader(state) {
+        return Object.prototype.hasOwnProperty.call(state.headers, 'Authorization');
+    },
 };
